fix: oracle types, try parse to query string and alias on getallcharachters

+++ PersonagemApi/DTOs/PersonagemDTO.cs
@@ -2,5 +2,7 @@
 {
     // buscar todos os personagens da base
     // buscar todos os personagens com filtros de códigos ou de nome
-    public record PersonagemDTO(int CdPersonagem, int CdClasse, string NmPersonagem, string? NrNivel, DateTime? DtCadastro, string? NmClasse);
+
+    // tive que colocar como decimal pq o dapper não acha o construtor com o tipo int (oracle mapeia para system.decimal mas dapper exige que bata os tipos)
+    public record PersonagemDTO(decimal CdPersonagem, decimal CdClasse, string NmPersonagem, string? NrNivel, DateTime? DtCadastro, string? NmClasse);
 }
